Move color scheme information into atoms
