+++ src/speedscope.tsx
@@ -1,7 +1,5 @@
 import {h, render} from 'preact'
-import {createAppStore} from './store'
 import {ApplicationContainer} from './views/application-container'
-import {Provider} from './lib/preact-redux'
 import {ThemeProvider} from './views/themes/theme'
 
 console.log(`speedscope v${require('../package.json').version}`)
@@ -15,16 +13,10 @@ if (module.hot) {
   module.hot.accept()
 }
 
-const lastStore: any = (window as any)['store']
-const store = lastStore ? createAppStore(lastStore.getState()) : createAppStore()
-;(window as any)['store'] = store
-
 render(
-  <Provider store={store}>
-    <ThemeProvider>
-      <ApplicationContainer />
-    </ThemeProvider>
-  </Provider>,
+  <ThemeProvider>
+    <ApplicationContainer />
+  </ThemeProvider>,
   document.body,
   document.body.lastElementChild || undefined,
 )
