--- src/store/actions.ts
@@ -1,6 +1,0 @@
-import {actionCreator} from '../lib/typed-redux'
-import {ColorScheme} from '.'
-
-export namespace actions {
-  export const setColorScheme = actionCreator<ColorScheme>('setColorScheme')
-}
